+++ managers/oclcCatalog.py
@@ -12,17 +12,14 @@ logger = createLog(__name__)
 class OCLCCatalogManager:
     CATALOG_URL = 'http://www.worldcat.org/webservices/catalog/content/{}?wskey={}'
     OCLC_SEARCH_URL = 'https://americas.discovery.api.oclc.org/worldcat/search/v2/'
-    # TODO: don't have init param oclcNo
-    def __init__(self, oclcNo):
-        self.oclcNo = oclcNo
+    def __init__(self):
         self.oclcKey = os.environ['OCLC_API_KEY']
         self.attempts = 0
 
-    # TODO: pass in oclcNo
-    def queryCatalog(self):
+    def queryCatalog(self, oclcNo):
         classifyResp = None
         self.attempts += 1
-        catalogQuery = self.CATALOG_URL.format(self.oclcNo, self.oclcKey)
+        catalogQuery = self.CATALOG_URL.format(oclcNo, self.oclcKey)
         if self.attempts > 3:
             return classifyResp
 
@@ -30,7 +27,7 @@ class OCLCCatalogManager:
             classifyResp = requests.get(catalogQuery, timeout=3)
         except (Timeout, ConnectionError):
             logger.warn('Failed to query URL {}'.format(catalogQuery))
-            return self.queryCatalog()
+            return self.queryCatalog(oclcNo)
 
         if classifyResp.status_code != 200:
             logger.warn('OCLC Catalog Request failed with status {}'.format(